% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenProcessing.R
\name{screenProcessing}
\alias{screenProcessing}
\title{Title}
\usage{
screenProcessing<-function(inputfile,controlStart,controlEnd,resultfile,maxsgRNA,minReadCount,zscore,orderOutput=TRUE,shortOutput=TRUE){
  
}
\arguments{
\item{inputfile}{Input file where each column represent a clone and each row a sgRNA. The first  Has to be a csv file (comma separated)}

\item{controlStart}{The colname of the first }

\item{controlEnd}{..}

\item{resultfile}{..}

\item{maxsgRNA}{..}

\item{minReadCount}{..}

\item{zscore}{..}

\item{orderOutput}{..}

\item{shortOutput}{..}
}
\value{
resultfile
}
\description{
Title
}
