% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenProcessing.R
\name{screenProcessing}
\alias{screenProcessing}
\title{Title}
\usage{
screenProcessing(inputfile, controlStart, controlEnd, resultfile, maxsgRNA,
  minReadCount, zscore)
}
\arguments{
\item{inputfile}{csv file (comma separated) where each column represent a clones and and each row.}

\item{controlStart}{..}

\item{controlEnd}{..}

\item{resultfile}{..}

\item{maxsgRNA}{..}

\item{minReadCount}{..}

\item{zscore}{..}
}
\value{
resultfile
}
\description{
Title
}
